Made ever greater stride with Terrain geenration and finally solved that heinous dicontinuity bug. Also fixed some mixcallanious bugs here and there

+++ Source/Noise.cpp
@@ -2,12 +2,12 @@
 
 NAMESPACE {
 
-  f32 Noise::fractal(function<f32(Vec2f,i32)> noise_func,
-		     Vec2f input, i32 num_octaves,
-		     f32 lac, f32 gain) {
-    f32 f = 1.0f;
-    f32 a = 1.0f;
-    f32 ret = 0;
+  f64 Noise::fractal(function<f64(Vec2d,i32)> noise_func,
+		     Vec2d input, i32 num_octaves,
+		     f64 lac, f64 gain) {
+    f64 f = 1.0f;
+    f64 a = 1.0f;
+    f64 ret = 0;
     for (i32 i=0; i<num_octaves; ++i) {
       ret += a*(noise_func(input*f,i));
       f *= lac;
@@ -24,15 +24,15 @@ NAMESPACE {
   }
 
   const static Vec2d PERLIN_OFFSET(1000000.0,2000000.0);
-  f32 FrankNoise::getValue(Vec2f point, i32 index) {
+  f64 FrankNoise::getValue(Vec2d point, i32 index) {
     Mat2d r = rot;
-    for (i32 i=0; i<index; ++i) {
+    for (i64 i=0; i<index; ++i) {
       r *= rot;
     }
     return perlin.getValue(rot*Vec2d(point) + PERLIN_OFFSET*(index+1));
   }
   
-    f32 FrankNoise::getPositive(Vec2f point, i32 index) {
+    f64 FrankNoise::getPositive(Vec2d point, i32 index) {
       return 0.5 + 0.5*getValue(point, index);
   }
 }
