+++ Source/Noise.hpp
@@ -8,9 +8,9 @@ NAMESPACE {
 
   struct Noise {
     
-    static f32 fractal(function<f32(Vec2f,i32)> noise_func,
-		       Vec2f input, i32 num_octaves,
-		       f32 lac = 2.0, f32 gain = 0.5);
+    static f64 fractal(function<f64(Vec2d,i32)> noise_func,
+		       Vec2d input, i32 num_octaves,
+		       f64 lac = 2.0, f64 gain = 0.5);
 
   };
 
@@ -20,9 +20,9 @@ NAMESPACE {
     Mat2d rot;
 
     FrankNoise();
-    f32 getValue(Vec2f point, i32 index = 0);
+    f64 getValue(Vec2d point, i32 index = 0);
     //equivalent to 0.5 + 0.5*(getValue(point, index))
-    f32 getPositive(Vec2f point, i32 index = 0);
+    f64 getPositive(Vec2d point, i32 index = 0);
     
   };
